% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCalculations.R
\name{cppHaversine}
\alias{cppHaversine}
\title{cpp haversine}
\usage{
cppHaversine(latFrom, lonFrom, latTo, lonTo, r = earthsRadius(),
  tolerance = 1e+09)
}
\arguments{
\item{latFrom}{latitude}

\item{lonFrom}{longitude}

\item{latTo}{latitude}

\item{lonTo}{latitude}

\item{r}{radius of earth}
}
\value{
distance in metres
}
\description{
Calculates the Haversine distance between two points
}
\examples{
dt1 <- data.table(lat1 = seq(-38, -37, by = 0.1),
  lon1 = seq(144, 145, by = 0.1),
  lat2 = seq(-35, -34, by = 0.1),
  lon2 = seq(145, 146, by = 0.1))

dt1[, distance := cppHaversine(lat1, lon1, lat2, lon2)]

}
