% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCalculations.R
\name{cppMidpoint}
\alias{cppMidpoint}
\title{cpp midpoint}
\usage{
cppMidpoint(latFrom, lonFrom, latTo, lonTo)
}
\arguments{
\item{latFrom}{latitude from}

\item{lonFrom}{longitude from}

\item{latTo}{latitude to}

\item{lonTo}{longitude to}
}
\value{
list of the latitude & longitude coordinates at the midpoint along the line
}
\description{
Finds the half-way point along the great circle path between two points
}
\examples{
dtMidpoint(0,0,0,52)
cppMidpoint(0,0,0,52)

dtMidpoint(25, 0, -25, 0)
cppMidpoint(25, 0, -25, 0)

dtMidpoint(-37,144,-38,145)
cppMidpoint(-37,144,-38,145)

dtMidpoint(25, 0, 35, 0)
dt <- data.table::data.table(lat = 25, lon = 0, lat2 = 35, lon2 = 0)
dt[, c("latMid", "lonMid") := dtMidpoint(lat, lon, lat2, lon2)]

cppMidpoint(25, 0, 35, 0)
dt <- data.table::data.table(lat = 25, lon = 0, lat2 = 35, lon2 = 0)
dt[, c("latMid", "lonMid") := cppMidpoint(lat, lon, lat2, lon2)]


}
