% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{dtAlongTrackDistance}
\alias{dtAlongTrackDistance}
\title{Along Track Distance}
\usage{
dtAlongTrackDistance(latFrom, lonFrom, latTo, lonTo, pointLat, pointLon,
  tolerance = 1e+09, r = earthsRadius())
}
\arguments{
\item{r}{}
}
\description{
Calculates the distance along a great-circle to the cross-track point
}
\seealso{
dtDist2gc
}

