% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCalculations.R
\name{dtDestination}
\alias{dtDestination}
\title{dt destination}
\usage{
dtDestination(latFrom, lonFrom, distance, bearing, r = earthsRadius())
}
\arguments{
\item{latFrom}{latitude from (in degrees)}

\item{lonFrom}{longitude from (in degrees)}

\item{distance}{distance in metres}

\item{bearing}{from north (in degrees)}

\item{r}{radius of earth in metres}
}
\value{
list of latitude and longitude coordinates at the destination
}
\description{
Calculates the destination coordinates from a starting point, initial bearing and distance travelling along a (shortest distance) great circle arc
}
\examples{

dtDestination(0, 0, earthsRadius(), 90)

dt <- data.table::data.table(lat = 0, lon = 0)
dt[, c("destinationLat", "destinationLon") := dtDestination(lat, lon, earthsRadius(), 90)]

}

