% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{dtDist2gc}
\alias{dtDist2gc}
\title{Dist 2 gc}
\usage{
dtDist2gc(latFrom, lonFrom, latTo, lonTo, pointLat, pointLon,
  tolerance = 1e+09, r = earthsRadius())
}
\arguments{
\item{r}{}
}
\description{
Finds the distance of a point to a great-circle path
}
\details{
Sometimes called 'cross-track' distance
}

