% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{dtEuclidean}
\alias{dtEuclidean}
\title{dt Euclidean}
\usage{
dtEuclidean(latFrom, lonFrom, latTo, lonTo)
}
\arguments{
\item{latFrom}{latitude from}

\item{lonFrom}{longitude from}

\item{latTo}{latitude to}

\item{lonTo}{longitude to}
}
\value{
Euclidean distance
}
\description{
Computes the Euclidean distance between two pairs of latitude / longitude coordinates
}

