% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCalculations.R
\name{dtHaversine}
\alias{dtHaversine}
\title{dt haversine
Calculates the Haversine distance between two points}
\usage{
dtHaversine(latFrom, lonFrom, latTo, lonTo, r = earthsRadius())
}
\arguments{
\item{latFrom}{latitude}

\item{lonFrom}{longitude}

\item{latTo}{latitude}

\item{lonTo}{latitude}

\item{r}{radius of earth}
}
\value{
distance in metres
}
\description{
dt haversine
Calculates the Haversine distance between two points
}
\examples{
dt <- data.table(lat1 = seq(-38, -37, by = 0.1),
  lon1 = seq(144, 145, by = 0.1),
  lat2 = seq(-35, -34, by = 0.1),
  lon2 = seq(145, 146, by = 0.1))

dt[, distance := dt.haversine(lat1, lon1, lat2, lon2)]

}

