% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{NearestPoints}
\alias{NearestPoints}
\title{Nearest Points}
\usage{
NearestPoints(x1, y1, x2, y2, tolerance = 1e+09, r = earthsRadius())
}
\arguments{
\item{r}{}
}
\description{
find nearest points between two vectors
}

