% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyline.R
\name{SimplifyPolyline}
\alias{SimplifyPolyline}
\title{Simplify Polyline}
\usage{
SimplifyPolyline(polyline, distanceTolerance = 1000)
}
\arguments{
\item{polyline}{encoded string of a polyline}

\item{distanceTolerance}{in metres}
}
\description{
simplifies a polyline using a vector-clustering algorithm
}

